reformat code

+++ blog/admin.py
@@ -21,30 +21,38 @@ class PictureAdminInline(admin.TabularInline):
 
 @admin.register(Category)
 class CategoryAdmin(admin.ModelAdmin):
-    fields = ['name']
-    list_display = ['id', 'name', 'numArticle']
-    list_editable = ['name']
+    fields = ['name', 'group']
+    list_display = ['id', 'name', 'group', 'timeModify', 'numArticle']
+    list_editable = ['name', 'group']
     inlines = [ArticleAdminInline]
 
+    def save_model(self, request, obj, form, change):
+        return super().save_related(request, obj, form, change)
+
     def save_related(self, request, form, formsets, change):
-        formset = formsets[0]
-        instances = formset.save(commit=False)
-        for obj in formset.deleted_objects:
-            obj.delete()
-        for instance in instances:
-            instance.author = instance.author or request.user
-            instance.modifier = request.user
-            instance.save()
-        formsets = formsets[1:]
+        if formsets:
+            formset = formsets[0]
+            instances = formset.save(commit=False)
+            for obj in formset.deleted_objects:
+                obj.delete()
+            for instance in instances:
+                instance.author = instance.author or request.user
+                instance.modifier = request.user
+                instance.save()
+            formsets = formsets[1:]
         super().save_related(request, form, formsets, change)
+
+    def delete_queryset(self, request, queryset):
+        for obj in queryset.all():
+            obj.delete()
 
 
 @admin.register(Article)
 class ArticleAdmin(admin.ModelAdmin):
     fields = ['title', 'category', 'MarkdownBody', 'timePublish']
-    list_display = ['title', 'category', 'hasPublished', 'timePublish', 'modifier', 'timeModify', 'author', 'timeCreate']
+    list_display = ['title', 'category', 'hasPublished', 'timePublish', 'modifier', 'timeModify']
     list_editable = ['category']
-    list_filter = ('author', 'modifier', 'category__name', 'timePublish')
+    list_filter = ['modifier', 'category__name', 'timePublish']
     inlines = [PictureAdminInline]
 
     def save_model(self, request, obj, form, change):
@@ -53,14 +61,15 @@ class ArticleAdmin(admin.ModelAdmin):
         obj.save()
 
     def save_related(self, request, form, formsets, change):
-        formset = formsets[0]
-        instances = formset.save(commit=False)
-        for obj in formset.deleted_objects:
-            obj.delete()
-        for instance in instances:
-            instance.painter = request.user
-            instance.save()
-        formsets = formsets[1:]
+        if formsets:
+            formset = formsets[0]
+            instances = formset.save(commit=False)
+            for obj in formset.deleted_objects:
+                obj.delete()
+            for instance in instances:
+                instance.painter = request.user
+                instance.save()
+            formsets = formsets[1:]
         super().save_related(request, form, formsets, change)
 
     def delete_queryset(self, request, queryset):
@@ -75,6 +84,14 @@ class PictureAdmin(admin.ModelAdmin):
     list_editable = ['article']
     list_filter = ['article__category__name', 'painter', 'timeCreate']
     readonly_fields = ['url']
+
+    def save_model(self, request, obj, form, change):
+        obj.painter = obj.painter or request.user
+        obj.save()
+
+    def delete_queryset(self, request, queryset):
+        for obj in queryset.all():
+            obj.delete()
 
 
 admin.site.site_header = settings.SITE_NAME
